constructor args type: unsigned int => Long

+++ src/main/kotlin/net/thiki/rest/sample/biz/issue/Issue.kt
@@ -1,15 +1,15 @@
 package net.thiki.rest.sample.biz.issue
 
-class Issue(val id: Long, val key: String, var reporter: Int, var summary: String) {
+class Issue(val id: Long, val key: String, var reporter: Long, var summary: String) {
 
     var description: String? = null
-    var assignee: Int? = 0
-    var status: Int? = 0
+    var assignee: Long? = 0
+    var status: Long? = 0
 
     /**
      * constructor used by mybatis
      */
-    constructor(id: Long, key: String, reporter: Int, summary: String, description: String?, assignee: Int?, status: Int?): this(id, key, reporter, summary){
+    constructor(id: Long, key: String, reporter: Long, summary: String, description: String?, assignee: Long?, status: Long?): this(id, key, reporter, summary){
         this.description = description
         this.assignee = assignee
         this.status = status
